+++ test/all_tests.dart
@@ -2,11 +2,13 @@ library all_tests;
 
 import 'package:unittest/html_enhanced_config.dart';
 
-import 'src/simple_event_bus_test.dart' as simpleEventbus;
+import 'event_bus_test.dart' as eventBusTest;
+import 'hierarchical_event_bus_test.dart' as hierarchicalEventBusTest;
 
 main() {
   useHtmlEnhancedConfiguration();
   
-  simpleEventbus.main();
+  eventBusTest.main();
+  hierarchicalEventBusTest.main();
 }