+++ test/event_bus_test.dart
@@ -1,0 +1,106 @@
+library event_bus_test;
+
+import 'dart:async';
+import 'package:unittest/unittest.dart';
+
+import 'package:event_bus/event_bus.dart';
+
+class EventA {
+  String text;
+  
+  EventA(this.text);
+}
+
+class EventB {
+  String text;
+  
+  EventB(this.text);
+}
+
+main() {
+  
+group('[EventBus]', () {
+
+  test('Fire one event', () {
+    // given
+    EventBus eventBus = new EventBus();
+    Future f = eventBus.on(EventA).toList();
+    
+    // when
+    eventBus.fire(new EventA('a1'));
+    eventBus.destroy();
+    
+    // then
+    return f.then((List events) {
+      expect(events.length, 1);
+    });
+  });
+  
+  test('Fire two events of same type', () {
+    // given
+    EventBus eventBus = new EventBus();
+    Future f = eventBus.on(EventA).toList();
+    
+    // when
+    eventBus.fire(new EventA('a1'));
+    eventBus.fire(new EventA('a2'));
+    eventBus.destroy();
+    
+    // then
+    return f.then((List events) {
+      expect(events.length, 2);
+    });
+  });
+  
+  test('Fire events of different type', () {
+    // given
+    EventBus eventBus = new EventBus();
+    Future f1 = eventBus.on(EventA).toList();
+    Future f2 = eventBus.on(EventB).toList();
+    
+    // when
+    eventBus.fire(new EventA('a1'));
+    eventBus.fire(new EventB('b1'));
+    eventBus.destroy();
+    
+    // then
+    return Future.wait(
+        [
+        f1.then((List events) {
+          expect(events.length, 1);
+        }),
+        f2.then((List events) {
+          expect(events.length, 1);
+        })
+        ]);
+  });
+  
+  test('Fire events of different type, receive all types', () {
+    // given
+    EventBus eventBus = new EventBus();
+    Future f = eventBus.on().toList();
+    
+    // when
+    eventBus.fire(new EventA('a1'));
+    eventBus.fire(new EventB('b1'));
+    eventBus.fire(new EventB('b2'));
+    eventBus.destroy();
+    
+    // then
+    return f.then((List events) {
+      expect(events.length, 3);
+    });
+  });
+});
+}
+
+
+
+
+
+
+
+
+
+
+